+++ .gitignore
@@ -1,7 +1,7 @@
 /build
 /dist
 /*.spec
-/__pycache__
+__pycache__
 *.exe
 *.bytes
 /.venv
